+++ ui/components/modals/RELEASENOTES.md
@@ -18,6 +18,7 @@
 - Changed assistive text for close buttons in modals, and added notes to accessibility documentation.
 - Changed default header level from `h2` to `h1` for semantics and added more description around using headers in the Accessibility section.
 - Removed redundant `title` attribute from modal close button.
+- Updated use of `aria-describedby` attribute for modal content, and clarified in accessibility documentation section.
 
 ## 2.14.0